New memory usage reduction, object oriented mode etc.
